Remove WebRTC provider
