--- packages/collaboration-extension/schema/provider.json
@@ -1,25 +1,0 @@
-{
-  "title": "Real-Time Collaboration Provider",
-  "description": "Provider for real-time collaboration (WebSocket or WebRTC).",
-  "jupyter.lab.setting-registry": {
-    "plugin": "@jupyter/collaboration-extension:provider"
-  },
-  "properties": {
-    "provider": {
-      "type": "string",
-      "enum": ["websocket", "webrtc"],
-      "default": "websocket",
-      "description": "The provider type for real-time collaboration. 'websocket' uses WebSocket, 'webrtc' uses WebRTC."
-    },
-    "signalingUrls": {
-      "type": "array",
-      "items": {
-        "type": "string"
-      },
-      "default": [],
-      "description": "List of signaling server URLs for WebRTC. If empty, defaults to the server's WebSocket URL with path 'api/signaling'."
-    }
-  },
-  "additionalProperties": false,
-  "type": "object"
-}
