+++ packages/collaboration-extension/src/index.ts
@@ -15,10 +15,6 @@ import {
   userEditorCursors
 } from './collaboration';
 import { sharedLink } from './sharedlink';
-import {
-  documentProviderFactoryPlugin,
-  awarenessProviderFactoryPlugin
-} from './provider';
 
 /**
  * Export the plugins as default.
@@ -29,9 +25,7 @@ const plugins: JupyterFrontEndPlugin<any>[] = [
   rtcGlobalAwarenessPlugin,
   rtcPanelPlugin,
   sharedLink,
-  userEditorCursors,
-  documentProviderFactoryPlugin,
-  awarenessProviderFactoryPlugin
+  userEditorCursors
 ];
 
 export default plugins;
