+++ packages/collaborative-drive/src/tokens.ts
@@ -22,7 +22,7 @@ export const ICollaborativeContentProvider =
 /**
  * The global awareness token.
  */
-export const IGlobalAwareness: Token<IAwareness> = new Token<IAwareness>(
+export const IGlobalAwareness = new Token<IAwareness>(
   '@jupyter/collaboration:IGlobalAwareness'
 );
