+++ packages/docprovider-extension/babel.config.js
@@ -3,6 +3,4 @@
  * Distributed under the terms of the Modified BSD License.
  */
 
-declare namespace NodeJS {
-  interface Timeout {}
-}
+module.exports = require('@jupyterlab/testing/lib/babel-config');
