+++ packages/docprovider-extension/package.json
@@ -48,6 +48,10 @@
     "clean:labextension": "rimraf ../../projects/jupyter-docprovider/jupyter_docprovider/labextension",
     "clean:all": "jlpm run clean:lib && jlpm run clean:labextension",
     "install:extension": "jlpm run build",
+    "test": "jest",
+    "test:cov": "jest --collect-coverage",
+    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
+    "test:debug:watch": "node --inspect-brk node_modules/.bin/jest --runInBand --watch",
     "watch": "run-p watch:src watch:labextension",
     "watch:src": "tsc -w",
     "watch:labextension": "jupyter labextension watch ."
@@ -72,7 +76,11 @@
   },
   "devDependencies": {
     "@jupyterlab/builder": "^4.5.0",
+    "@jupyterlab/testing": "^4.5.0",
+    "@jupyterlab/testutils": "^4.5.0",
+    "@types/jest": "^29.2.0",
     "@types/react": "~18.3.1",
+    "jest": "^29.5.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^4.1.2",
     "typescript": "~5.1.6"
