+++ packages/docprovider-extension/tsconfig.test.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.test.json",
+  "include": ["src/*", "src/__tests__/*"]
+}
