+++ packages/docprovider/src/index.ts
@@ -7,11 +7,9 @@
  * @module docprovider
  */
 
-export * from './awareness';
 export * from './notebookCellExecutor';
 export * from './requests';
 export * from './ydrive';
-export * from './yprovider';
 export * from './TimelineSlider';
 export * from './tokens';
 export * from './forkManager';
