--- packages/docprovider/src/typings.d.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) Jupyter Development Team.
- * Distributed under the terms of the Modified BSD License.
- */
-
-declare namespace NodeJS {
-  interface Timeout {}
-}
