--- packages/y-webrtc/.dockerignore
@@ -1,4 +1,0 @@
-Dockerfile
-.dockerignore
-node_modules
-.git
