--- packages/y-webrtc/.gitignore
@@ -1,3 +1,0 @@
-dist
-node_modules
-.vscode
