--- packages/y-webrtc/README.md
@@ -1,3 +1,0 @@
-# @jupyter/y-webrtc
-
-A JupyterLab package which provides a WebRTC connector for [Yjs](https://github.com/yjs/yjs). This is a fork of y-webrtc at [this commit](https://github.com/yjs/y-webrtc/tree/c411f1d7223b68f3c6fc9c5901a19819a17db667).
