--- packages/y-webrtc/index.html
@@ -1,14 +1,0 @@
-<!--
-  ~ Copyright (c) Jupyter Development Team.
-  ~ Distributed under the terms of the Modified BSD License.
--->
-
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Testing y-webrtc</title>
-</head>
-<body>
-  <script type="module" src="./dist/test.js"></script>
-</body>
-</html>
