--- packages/y-webrtc/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "@jupyter/y-webrtc",
-  "version": "10.3.0",
-  "description": "WebRTC provider for Yjs",
-  "type": "module",
-  "main": "./lib/y-webrtc.js",
-  "types": "./lib/y-webrtc.d.ts",
-  "sideEffects": false,
-  "funding": {
-    "type": "GitHub Sponsors ❤",
-    "url": "https://github.com/sponsors/dmonad"
-  },
-  "exports": {
-    ".": {
-      "import": "./lib/y-webrtc.js",
-      "types": "./lib/y-webrtc.d.ts"
-    }
-  },
-  "scripts": {
-    "build": "tsc -b",
-    "install:extension": "jlpm run build",
-    "clean": "rimraf lib tsconfig.tsbuildinfo",
-    "watch": "tsc -b --watch",
-    "test": "npm run lint",
-    "lint": "tsc"
-  },
-  "bin": {
-    "y-webrtc-signaling": "./bin/server.js"
-  },
-  "files": [
-    "lib/**/*.{d.ts,js,js.map}",
-    "bin/*",
-    "schema/*"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/yjs/y-webrtc.git"
-  },
-  "keywords": [
-    "Yjs"
-  ],
-  "author": "Kevin Jahns <kevin.jahns@protonmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/yjs/y-webrtc/issues"
-  },
-  "homepage": "https://github.com/yjs/y-webrtc#readme",
-  "standard": {
-    "ignore": [
-      "/dist",
-      "/node_modules"
-    ]
-  },
-  "dependencies": {
-    "lib0": "^0.2.42",
-    "simple-peer": "^9.11.0"
-  },
-  "devDependencies": {
-    "@types/simple-peer": "^9.11.3",
-    "concurrently": "^5.3.0",
-    "http-server": "^0.12.3",
-    "standard": "^14.3.4",
-    "typescript": "~5.1.6",
-    "y-protocols": "^1.0.6",
-    "yjs": "^13.6.8"
-  },
-  "peerDependencies": {
-    "y-protocols": "^1.0.6",
-    "yjs": "^13.6.8"
-  },
-  "optionalDependencies": {
-    "ws": "^8.14.2"
-  },
-  "engines": {
-    "node": ">=12"
-  }
-}
