--- packages/y-webrtc/test/index.js
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) Jupyter Development Team.
- * Distributed under the terms of the Modified BSD License.
- */
-
-import * as crypto from './crypto.test.js'
-
-import { runTests } from 'lib0/testing'
-import { isBrowser, isNode } from 'lib0/environment'
-import * as log from 'lib0/logging'
-
-if (isBrowser) {
-  log.createVConsole(document.body)
-}
-runTests({
-  crypto
-}).then(success => {
-  /* istanbul ignore next */
-  if (isNode) {
-    process.exit(success ? 0 : 1)
-  }
-})
