--- packages/y-webrtc/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "declaration": true,
-    "skipLibCheck": true
-  },
-  "include": ["./src/**/*.ts"],
-  "exclude": ["./node_modules/**", "./lib/**"]
-}
