--- projects/jupyter-server-signaling/README.md
@@ -1,3 +1,0 @@
-# jupyter-server-signaling
-
-jupyter-server extension for WebRTC signaling.
