--- projects/jupyter-server-signaling/jupyter-config/jupyter_server_signaling.json
@@ -1,7 +1,0 @@
-{
-  "ServerApp": {
-    "jpserver_extensions": {
-      "jupyter_server_signaling": true
-    }
-  }
-}
