--- projects/jupyter-server-signaling/pyproject.toml
@@ -1,43 +1,0 @@
-# Copyright (c) Jupyter Development Team.
-# Distributed under the terms of the Modified BSD License.
-
-[build-system]
-build-backend = "hatchling.build"
-requires = ["hatchling>=1.4.0"]
-
-[project]
-name = "jupyter-server-signaling"
-version = "0.2.1"
-readme = "README.md"
-license = { file = "LICENSE" }
-requires-python = ">=3.9"
-description = "jupyter-server extension for WebRTC signaling."
-classifiers = [
-    "Intended Audience :: Developers",
-    "Intended Audience :: System Administrators",
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: BSD License",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Framework :: Jupyter"
-]
-authors = [
-    { name = "Jupyter Development Team", email = "jupyter@googlegroups.com" },
-]
-dependencies = [
-    "jupyter_server>=2.4.0,<3.0.0",
-    "websockets",
-]
-
-[tool.hatch.build.targets.sdist]
-exclude = ["/.github", "/binder", "node_modules"]
-
-[tool.hatch.build.targets.wheel.shared-data]
-"jupyter-config/jupyter_server_signaling.json" = "etc/jupyter/jupyter_server_config.d/jupyter_server_signaling.json"
-
-[tool.check-wheel-contents]
-ignore = ["W002"]
