+++ scripts/dev_install.py
@@ -11,18 +11,11 @@ def execute(cmd: str, cwd: Optional[Path] = None) -> None:
 
 
 def install_dev() -> None:
-    install_build_deps = (
-        "python -m pip install jupyterlab>=4.4.0,<5 hatchling hatch-jupyter-builder editables"
-    )
+    install_build_deps = "python -m pip install jupyterlab>=4.4.0,<5"
     install_js_deps = "jlpm install"
 
     python_package_prefix = "projects"
-    python_packages = [
-        "jupyter-collaboration-ui",
-        "jupyter-docprovider",
-        "jupyter-server-ydoc",
-        "jupyter-server-signaling",
-    ]
+    python_packages = ["jupyter-collaboration-ui", "jupyter-docprovider", "jupyter-server-ydoc"]
 
     execute(install_build_deps)
     execute(install_js_deps)
@@ -30,10 +23,10 @@ def install_dev() -> None:
     for py_package in python_packages:
         real_package_name = py_package.replace("-", "_")
         execute(f"pip uninstall {real_package_name} -y")
-        execute(f"pip install -e {python_package_prefix}/{py_package}[test] --no-build-isolation")
+        execute(f"pip install -e {python_package_prefix}/{py_package}[test]")
 
         # List of server extensions
-        if py_package in ["jupyter-server-ydoc", "jupyter-server-signaling"]:
+        if py_package in ["jupyter-server-ydoc"]:
             execute(f"jupyter server extension enable {real_package_name}")
 
         # List of jupyterlab extensions
