+++ tsconfig.json
@@ -13,11 +13,11 @@
     "noUnusedLocals": true,
     "preserveWatchOutput": true,
     "resolveJsonModule": true,
+    "skipLibCheck": true,
     "strict": true,
     "strictNullChecks": true,
     "target": "es2018",
     "types": [],
-    "skipLibCheck": true,
     "lib": ["DOM", "ES2018", "ES2020.Intl"]
   }
 }
